# outputs/deps-2026-09/openapi-v1.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RetentionPolicyUpdate": {
        "properties": {
          "policy": {
            "items": {
              "$ref": "#/components/schemas/RetentionRule"
            },
            "title": "Policy",
            "type": "array"
          }
        },
        "required": [
          "policy"
        ],
        "title": "RetentionPolicyUpdate",
        "type": "object"
      },
      "RetentionRule": {
        "properties": {
          "max_age_days": {
            "title": "Max Age Days",
            "type": "integer"
          },
          "resolution": {
            "title": "Resolution",
            "type": "string"
          }
        },
        "required": [
          "resolution",
          "max_age_days"
        ],
        "title": "RetentionRule",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Self-hosted stock analysis platform for Thai and US markets",
    "title": "StockViz API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/admin/retention-policy": {
      "get": {
        "description": "Get current data retention policy and disk usage stats.",
        "operationId": "get_retention_policy_api_v1_admin_retention_policy_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Retention Policy",
        "tags": [
          "admin"
        ]
      },
      "put": {
        "description": "Update data retention policy. Stored in Redis, read by housekeeping worker.",
        "operationId": "update_retention_policy_api_v1_admin_retention_policy_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RetentionPolicyUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Retention Policy",
        "tags": [
          "admin"
        ]
      }
    },
    "/api/v1/admin/retention-policy/run-now": {
      "post": {
        "description": "Trigger housekeeping worker immediately.",
        "operationId": "run_housekeeping_now_api_v1_admin_retention_policy_run_now_post",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Run Housekeeping Now",
        "tags": [
          "admin"
        ]
      }
    },
...
  }
}